Excepciones

+++ src/main/java/nomina/soft/backend/repositories/BoletaDePagoRepository.java
@@ -1,11 +1,10 @@
 package nomina.soft.backend.repositories;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import nomina.soft.backend.models.BoletaDePagoModel;
 
 @Repository
-public interface BoletaDePagoRepository extends JpaRepository<BoletaDePagoModel,Integer> {
+public interface BoletaDePagoRepository extends JpaRepository<BoletaDePagoModel,Long> {
         
 }
